refactor(evm): changes based on review feedbacks
