+++ apps/evm/src/constants/functionKey.ts
@@ -70,6 +70,7 @@ enum FunctionKey {
   GET_ACCOUNT_TRANSACTION_HISTORY = 'GET_ACCOUNT_TRANSACTION_HISTORY',
   GET_SIMULATED_POOL = 'GET_SIMULATED_POOL',
   GET_SWAP_QUOTE = 'GET_SWAP_QUOTE',
+  GET_PROPOSALS_COUNT_FROM_API = 'GET_PROPOSALS_COUNT_FROM_API',
 }
 
 export default FunctionKey;
