--- apps/evm/src/constants/landing.ts
@@ -1,11 +1,0 @@
-// TODO: update URLs
-
-export const DAPP_URL = 'https://app.venus.io';
-export const LANDING_PAGE_URL = 'https://venus.io';
-export const DOC_URL = 'https://docs-v4.venus.io';
-export const WHITEPAPERS_URL =
-  'https://github.com/VenusProtocol/venus-protocol-documentation/tree/main/whitepapers';
-
-export const DISCORD_SERVER_URL = 'https://discord.com/servers/venus-protocol-912811548651708448';
-
-export const COMMUNITY_URL = 'https://community.venus.io/';
