+++ apps/evm/src/constants/production.ts
@@ -1,3 +1,13 @@
 export const MAIN_PRODUCTION_HOST = 'app.venus.io';
 export const VENUS_X_URL = 'https://x.venus.io';
 export const E_MODE_DOC_URL = 'https://docs-v4.venus.io/whats-new/emode';
+
+export const DAPP_URL = 'https://app.venus.io';
+export const LANDING_PAGE_URL = 'https://venus.io';
+export const DOC_URL = 'https://docs-v4.venus.io';
+export const WHITEPAPERS_URL =
+  'https://github.com/VenusProtocol/venus-protocol-documentation/tree/main/whitepapers';
+
+export const DISCORD_SERVER_URL = 'https://discord.com/servers/venus-protocol-912811548651708448';
+
+export const COMMUNITY_URL = 'https://community.venus.io/';
