--- apps/evm/src/containers/Container/index.tsx
@@ -1,8 +1,0 @@
-import { cn } from 'components';
-
-const containerClassName = 'w-full px-4 sm:px-4 md:px-6 lg:px-12 lg:mx-auto';
-
-export const Container: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({
-  className,
-  ...props
-}) => <div className={cn(containerClassName, className)} {...props} />;
