+++ apps/evm/src/pages/Landing/Protection/index.tsx
@@ -1,5 +1,4 @@
 import { cn } from 'components';
-import { Container } from 'containers/Container';
 import { useTranslation } from 'libs/translations';
 import bugBountyImg from './assets/bugBounty.png';
 import protectionImg from './assets/protection.png';
@@ -8,9 +7,11 @@ interface IProtectionProps {
   className?: string;
 }
 
-const itemClassName =
-  'relative flex flex-1 flex-col min-h-105 justify-between p-6 pb-0 border border-solid border-lightGrey rounded-3xl sm:min-h-100 xl:min-h-139 xl:p-10 xl:pb-0';
-const bgImgClassName = 'w-full bottom-0';
+const itemClassName = cn(
+  'relative flex flex-1 flex-col min-h-105 justify-between p-6 pb-0 border border-solid border-lightGrey rounded-3xl sm:min-h-100 xl:min-h-139 xl:p-10 xl:pb-0',
+);
+
+const bgImgClassName = cn('w-full bottom-0');
 
 const textClassName = cn(
   '[&_h2]:text-[1.5rem] [&_h2]:font-semibold xl:[&_h2]:text-[2rem]',
@@ -21,20 +22,19 @@ export const Protection: React.FC<IProtectionProps> = ({ className }) => {
   const { t } = useTranslation();
 
   return (
-    <Container className={cn('mt-15 md:mt-20 lg:mt-25', textClassName, className)}>
+    <div className={cn('mt-15 md:mt-20 lg:mt-25', textClassName, className)}>
       <ul className={'flex flex-col justify-between gap-6 sm:flex-row xl:gap-8'}>
         <li
           className={cn(
             itemClassName,
             'bg-radial-[62.14%_57.90%_at_50.00%_50.00%,rgba(23,46,98,0.50)_0%,rgba(18,22,32,0.50)_100%]',
           )}
-          key="bounty"
         >
           <div className={'xl:mb-21'}>
             <h2 className="mb-4">{t('landing.protection.challengeTitle')}</h2>
             <p className="mb-8 md:mb-0">{t('landing.protection.challengeText')}</p>
           </div>
-          <img loading="lazy" className={cn(bgImgClassName)} src={bugBountyImg} alt="bugBounty" />
+          <img loading="lazy" className={bgImgClassName} src={bugBountyImg} alt="bugBounty" />
         </li>
         <li
           className={cn(
@@ -47,9 +47,9 @@ export const Protection: React.FC<IProtectionProps> = ({ className }) => {
             <h2 className="mb-4">{t('landing.protection.protectionTitle')}</h2>
             <p className="mb-8 md:mb-0">{t('landing.protection.protectionText')}</p>
           </div>
-          <img loading="lazy" className={cn(bgImgClassName)} src={protectionImg} alt="protection" />
+          <img loading="lazy" className={bgImgClassName} src={protectionImg} alt="protection" />
         </li>
       </ul>
-    </Container>
+    </div>
   );
 };
