+++ apps/evm/src/pages/Landing/Wallets/index.tsx
@@ -1,4 +1,3 @@
-import { Container } from 'containers/Container';
 import { useTranslation } from 'libs/translations';
 import binanceLogoSrc from './assets/binance.svg';
 import braveLogoSrc from './assets/brave.svg';
@@ -74,7 +73,7 @@ export const Wallets: React.FC = () => {
 
   return (
     <section className="my-15 md:my-20 xl:my-25">
-      <Container>
+      <div>
         <h2 className="text-[2rem] text-center mx-auto mb-10">{t('landing.wallets.title')}</h2>
 
         <div className="grid grid-cols-3 gap-4 sm:grid-cols-4 md:grid-cols-6 xl:gap-6">
@@ -84,7 +83,7 @@ export const Wallets: React.FC = () => {
             </div>
           ))}
         </div>
-      </Container>
+      </div>
     </section>
   );
 };
